+++ app/models/__init__.py
@@ -1,0 +1,3 @@
+from .profile import Profile
+from .track import Track, SavedTrack, Library
+from .user import User
