+++ app/models/track.py
@@ -1,10 +1,9 @@
-from spotify import api
-from spotify import utils
-from spotify import query_builder
+import spotify.api
+import spotify.utils
+import query_builder
 
-# Represents a Track Object
 class Track:
-
+    """ Represents a Spotify Track Object """
     RESOURCE_TYPE = 'track'
 
     # List of keys to filter by when getting simple json
@@ -13,38 +12,28 @@ class Track:
 
     @staticmethod
     def find_by_query(auth_token, search_query):
-        track_data = api.search(auth_token, search_query, Track.RESOURCE_TYPE)
-        if utils.successful_request(track_data):
-            print(track_data.keys())
-            return Track(auth_token, track_data['tracks']['items'][0])
-        else:
-            return None
+        track_data = spotify.api.search(auth_token, search_query, Track.RESOURCE_TYPE)
+        return Track(auth_token, track_data['tracks']['items'][0])
     
     @staticmethod
     def find_by_id(auth_token, _id):
-        track_data = api.get_track(auth_token, _id)
-        if utils.successful_request(track_data):
-            return Track(auth_token, track_data)
-        else:
-            return None
+        track_data = spotify.api.get_track(auth_token, _id)
+        return Track(auth_token, track_data)
     
     # Filter data to only include keys of interest
     def filter_track_data(self):
-        response = utils.filter_dict(self.track_info, Track.TRACK_KEYS)
+        response = spotify.utils.filter_dict(self.track_info, Track.TRACK_KEYS)
         # Artists is a list of artist objects, convert to comma separated string of artist names
         response['artists'] = ', '.join(list(map(lambda x: x['name'], response['artists'])))
         # Include audio feature keys if audio features have been populated
         if self.audio_features:
-            response.update({'audio_features': utils.filter_dict(self.audio_features, Track.AUDIO_FEATURE_KEYS)})
+            response.update({'audio_features': spotify.utils.filter_dict(self.audio_features, Track.AUDIO_FEATURE_KEYS)})
         return response
     
     # Gets Audio Analysis for this track
     def perform_audio_analysis(self):
-        response = api.track_audio_features(self._auth_token, self.track_info['id'])
-        if utils.successful_request(response):
-            self.audio_features = response
-            return True
-        return False
+        response = spotify.api.track_audio_features(self._auth_token, self.track_info['id'])
+        self.audio_features = response
         
     # Formats and returns important fields
     def to_simple_json(self):
@@ -73,29 +62,18 @@ class Track:
         self.audio_features = features_json
 
 
-# Represents a Profile Object
-class Profile:
-
-    # Gets user profile associated with auth token
-    def load_profile_info(self):
-        return api.get_current_profile(self._auth_token)
-
-    def __init__(self, auth_token):
-        self._auth_token = auth_token
-        self.profile_info = self.load_profile_info()
-
-# Represents a Track saved in a user's library
 class SavedTrack(Track):
+    """ Represents a Track saved in a user's library """
     def __init__(self, auth_token, saved_track_json, features_json = None):
         super().__init__(auth_token, saved_track_json.get('track'), features_json)
         self.added_at = saved_track_json.get('added_at')
 
 
-# Represents a User's Library
 class Library:
+    """ Represents a Collection of Tracks (ex. Library or Playlist) """
     # Hits Spotify endpoint and returns list of SavedTrack objects
     def get_library(self):
-        track_list = api.get_saved_tracks(self._auth_token).get('saved_tracks')
+        track_list = spotify.api.get_saved_tracks(self._auth_token).get('saved_tracks')
         return list(map(lambda obj: SavedTrack(self._auth_token, obj), track_list))
 
     # Filters the library based on a query (returns list of SavedTracks that match the query)
